JBIDE-13434 BrowserSim: add taking screenshot ability
menu simplification
