+++ plugins/org.jboss.tools.vpe.browsersim/src/org/jboss/tools/vpe/browsersim/ui/Messages.java
@@ -83,13 +83,10 @@ public class Messages {
 	public static String ManageDevicesDialog_USE_SKINS;
 	public static String ManageDevicesDialog_USER_AGENT;
 	public static String ManageDevicesDialog_WIDTH;
-	public static String ScreenshotDialog_Screenshot;
-	public static String ScreenshotDialog_Save;
-	public static String ScreenshotDialog_SaveAndOpen;
-	public static String ScreenshotDialog_SaveAs;
-	public static String ScreenshotDialog_SaveAsAndOpen;
-	public static String ScreenshotDialog_CopyPath;
-	public static String ScreenshotDialog_CopyImage;
+	public static String Screenshots_Screenshot;
+	public static String Screenshots_Save;
+	public static String Screenshots_SaveAs;
+	public static String Screenshots_CopyToClipboard;
 	public static String SizeWarningDialog_CANCEL;
 	public static String SizeWarningDialog_DESKTOP_SIZE_TOO_SMALL_HORIZONTAL;
 	public static String SizeWarningDialog_DESKTOP_SIZE_TOO_SMALL_VERTICAL;
