+++ plugins/org.jboss.tools.vpe.browsersim/src/org/jboss/tools/vpe/browsersim/ui/messages.properties
@@ -69,13 +69,10 @@ ManageDevicesDialog_LOAD_DEFAULTS=Load Defaults
 ManageDevicesDialog_USER_AGENT=User-Agent
 ManageDevicesDialog_USE_SKINS=Use skins
 ManageDevicesDialog_WIDTH=Width
-ScreenshotDialog_Screenshot=Screenshot
-ScreenshotDialog_Save=Save
-ScreenshotDialog_SaveAndOpen=Save and open
-ScreenshotDialog_SaveAs=Save as
-ScreenshotDialog_SaveAsAndOpen=Save as and open
-ScreenshotDialog_CopyPath=Copy path to clipboard
-ScreenshotDialog_CopyImage=Copy image to clipboard
+Screenshots_Screenshot=Screenshot
+Screenshots_Save=Save
+Screenshots_SaveAs=Save as...
+Screenshots_CopyToClipboard=Copy to clipboard
 SizeWarningDialog_DEVICE_SIZE_WILL_BE_TRUNCATED=Device size will be truncated
 SizeWarningDialog_DESKTOP_SIZE_TOO_SMALL_VERTICAL=Your desktop size ({0}x{1} pixels) is smaller than what {2} needs in vertical layout ({3}x{4}).\n\
 Device size will be truncated to fit your desktop.
